Move SimpleDeployCommandError to plugin_utils.
